requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
Pillow==9.1.0
